Change the relationship between store and address from (Many-to-Many) into (One-to-one), so each store has only one address. Edit the create-store-form to be able to: 1- ceate Address. 2-create Store and assign the address to it. 3-Assign Myself as an owner to the store. Edit delete Store to be able to delete the addres of the store after deleting the store itself. Delete the 3 components that were responsible for assigning addresses to the stores.
