+++ schemas/index.ts
@@ -45,14 +45,21 @@ export const SettingSchema = z.object({
         return true;
     }, {message: "password is required!", path: ["password"]});
 
-export const addStoreSchema = z.object( {
+export const createStoreSchema = z.object( {
     id: z.string().min(1),
     name: z.string().min(1,"Store Name"),
     department: z.enum([Department.Supermarket, Department.Meat, Department.Perfume]),
+    addressId: z.optional(z.string().min(1)),
+    city: z.string().min(1),
+    street_name: z.string().min(1),
+    house_number: z.string().min(1),
+    postal_code : z.string().min(1),
+    countryId: z.string().min(1),
 });
 
 export const deleteStoreSchema = z.object({
-    id: z.string().min(1)
+    id: z.string().min(1),
+    addressId: z.string().min(1)
 });
 
 export const assignSellerSchema = z.object({
@@ -60,12 +67,12 @@ export const assignSellerSchema = z.object({
     storeId: z.string().min(1)
 });
 
-export const createStoreAddressSchema = z.object({
-    id: z.optional(z.string().min(1)),
-    city: z.string().min(1),
-    street_name: z.string().min(1),
-    house_number: z.string().min(1),
-    postal_code : z.string().min(1),
-    countryId: z.string().min(1),
-    storeId: z.string().min(1),
-});
+// export const createStoreAddressSchema = z.object({
+//     id: z.optional(z.string().min(1)),
+//     city: z.string().min(1),
+//     street_name: z.string().min(1),
+//     house_number: z.string().min(1),
+//     postal_code : z.string().min(1),
+//     countryId: z.string().min(1),
+//     storeId: z.string().min(1),
+// });
